fix: crash when trying to take another survey after taking one
This changes the behaviour thought. Now after taking a survey instead of
going to `/(auth)/(tabs)/visits` (which is in another stack) we are
pop-ing all the way to the first page in the stack which is
`/(auth)/(visits)/select-house`. This seems to be related again to an
issue with expo-router, but there are fixes that we will only get in skd
53 but we will not update the sdk for now.