fix: refech meData on visits and fix non-existent value in the visit owner selection
The app was crashing when you changed the team of an user and switched to the visits page where the visit owner selector was render due to stale user data and fetching the wrong team. Now we're re-fetching the user date on visits page focus and also handling when you selected a member in the previous team that doesn't exist in the current team.