fix: allow logout even if the user object is not present
Weird bug that probably has to do with the hydration process and
persisted user. If you the access and refresh token expires and the app
was closed and you open it, the app will get in an infinite loop trying
to logout and not being able to because it expects a user object to be
present but probably the hydration failed.