feat: logout if the token is expired
I’ve tried refreshing the token but it seems that the refreshed access 
token doens’t expires like the first one. I believe there’s something 
wrong with our auth implementation in the backend.

For now, and since the past implementation was not redirecting when the refresh token expired, this is a win even though we’re not implementing the refresh token.